move discord services to discord service group
